refactor(web): structure du composant editbox
