+++ apps/web/design/app/organisms/_editbox-input.scss
@@ -39,9 +39,3 @@
 		--shadow-color: var(--user-bg-color);
 	}
 }
-
-.editbox {
-	button[type="submit"] svg {
-		transform: scale(0.75) translateX(2px);
-	}
-}
